Change filepath function locations
